streaming_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pika==1.3.2
websockets==12.0
python-multipart==0.0.6
